Toggle tree to specified directory (could be the CWD of the currently opened file) as root node if supplied. Sample mappings: nmap	,to	:NERDTreeToggle<CR> nmap	,tr	:exec 'NERDTreeToggle ' . <some_root_dir><CR> nmap	,tc	:exec 'NERDTreeToggle ' . getcwd()<CR>
,to will work as before
,tr and ,tc will switch the root node to the supplied directory

The behind-the-scene reason of this change is that if we use NERDTree or NERDTreeFind to open up a new root or to find the file not in the current tree, a new tree buffer is created.  This internally increases and wastes buf number and makes the buf explorer plugin (such as minibufexplorer) a bit awkard to use with jumping buf numbers and swithing buffers by typing buf number as I do. With this change it will stick to the existing tree buffer.  It suits my need this way.